run some calculations on cpu when the suset is small

+++ pennylane_lightning/core/src/simulators/lightning_tensor/tncuda/TNCudaBase.hpp
@@ -350,7 +350,8 @@ class TNCudaBase : public TensornetBase<PrecisionT, Derived> {
 
         std::vector<int32_t> projected_modes{};
 
-        for (int32_t idx = 0; idx < BaseType::getNumQubits(); idx++) {
+        for (int32_t idx = 0;
+             idx < static_cast<int32_t>(BaseType::getNumQubits()); idx++) {
             auto it = std::find(stateModes.begin(), stateModes.end(), idx);
             if (it == stateModes.end()) {
                 projected_modes.emplace_back(idx);
